ServiceManager project is now more generic and will install more than just HL7 service. Specifically, the changes were made so it could manage the OpenDentCustListener service also.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@10827 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ ServiceManager/FormServiceManage.cs
@@ -14,10 +14,17 @@ using System.IO;
 
 namespace ServiceManager {
 	public partial class FormServiceManage:Form {
+		private FileInfo _serviceFile {
+			get {
+				if(File.Exists(textPathToExe.Text)) {
+					return new FileInfo(textPathToExe.Text);
+				}
+				return null;
+			}
+		}
 		public ServiceController ServControllerCur;
 		public List<ServiceController> allOpenDentServices;
-		public string pathToExe;
-
+		
 		public FormServiceManage() {
 			InitializeComponent();
 		}
@@ -34,10 +41,7 @@ namespace ServiceManager {
 				textName.Text=ServControllerCur.ServiceName;
 				RegistryKey hklm=Registry.LocalMachine;
 				hklm=hklm.OpenSubKey(@"System\CurrentControlSet\Services\"+ServControllerCur.ServiceName);
-				pathToExe=hklm.GetValue("ImagePath").ToString();
-				pathToExe=pathToExe.Replace("\"","");
-				pathToExe=Path.GetDirectoryName(pathToExe);
-				textPathToExe.Text=pathToExe;
+				textPathToExe.Text=hklm.GetValue("ImagePath").ToString().Replace("\"","");
 				textStatus.Text="Installed";
 				butInstall.Enabled=false;
 				butUninstall.Enabled=true;
@@ -69,31 +73,33 @@ namespace ServiceManager {
 		}
 
 		private void butInstall_Click(object sender,EventArgs e) {
+			if(_serviceFile==null) {
+				MessageBox.Show("Select a valid service path");
+				return;
+			}
 			if(textName.Text.Length<8 || textName.Text.Substring(0,8)!="OpenDent") {
 				MessageBox.Show("Error.  Service name must begin with \"OpenDent\".");
 				return;
 			}
 			for(int i=0;i<allOpenDentServices.Count;i++) {//create list of all OpenDent service install paths to ensure only one service can be installed from each directory
 				if(textName.Text==allOpenDentServices[i].ServiceName) {
-					MessageBox.Show("Error.  An OpenDentHL7 service with this name is already installed.  Names must be unique.");
+					MessageBox.Show("Error.  A service with this name is already installed.  Names must be unique.");
 					return;
 				}
 				RegistryKey hklm=Registry.LocalMachine;
 				hklm=hklm.OpenSubKey(@"System\CurrentControlSet\Services\"+allOpenDentServices[i].ServiceName);
-				string installedServicePath=hklm.GetValue("ImagePath").ToString();
-				installedServicePath=installedServicePath.Replace("\"","");
-				installedServicePath=Path.GetDirectoryName(installedServicePath);
-				if(installedServicePath==textPathToExe.Text) {
-					MessageBox.Show("Error.  Cannot install an HL7 service from this directory.  Another OpenDentHL7 service is already installed from this directory.");
+				string installedServicePath=hklm.GetValue("ImagePath").ToString().Replace("\"","");
+				if(installedServicePath==_serviceFile.FullName) {
+					MessageBox.Show("Error.  Cannot install service.  This service is already installed from this directory.");
 					return;
 				}
 			}
 			Process process=new Process();
-			process.StartInfo.WorkingDirectory=textPathToExe.Text;
-			process.StartInfo.FileName="installutil.exe";
+			process.StartInfo.WorkingDirectory=_serviceFile.DirectoryName;
+			process.StartInfo.FileName=Path.Combine(Directory.GetCurrentDirectory(),"installutil.exe");
 			//new strategy for having control over servicename
 			//InstallUtil /ServiceName=OpenDentHL7_abc OpenDentHL7.exe
-			process.StartInfo.Arguments="/ServiceName="+textName.Text+" OpenDentHL7.exe";
+			process.StartInfo.Arguments="/ServiceName="+textName.Text+" "+_serviceFile.Name;
 			process.Start();
 			try {
 				process.WaitForExit(10000);
@@ -114,11 +120,15 @@ namespace ServiceManager {
 		}
 
 		private void butUninstall_Click(object sender,EventArgs e) {
+			if(_serviceFile==null) {
+				MessageBox.Show("Selected service has an invalid path");
+				return;
+			}
 			RegistryKey hklm=Registry.LocalMachine;
 			Process process=new Process();
-			process.StartInfo.WorkingDirectory=textPathToExe.Text;
-			process.StartInfo.FileName="installutil.exe";
-			process.StartInfo.Arguments="/u /ServiceName="+textName.Text+" OpenDentHL7.exe";
+			process.StartInfo.WorkingDirectory=_serviceFile.DirectoryName;
+			process.StartInfo.FileName=Path.Combine(Directory.GetCurrentDirectory(),"installutil.exe");
+			process.StartInfo.Arguments="/u /ServiceName="+textName.Text+" "+_serviceFile.Name;
 			process.Start();
 			try {
 				process.WaitForExit(10000);
@@ -170,13 +180,17 @@ namespace ServiceManager {
 		private void butRefresh_Click(object sender,EventArgs e) {
 			FormServiceManager_Load(this,e);
 		}
-
+				
 		private void butBrowse_Click(object sender,EventArgs e) {
-			FolderBrowserDialog dlg=new FolderBrowserDialog();
-			dlg.SelectedPath=textPathToExe.Text;
-			if(dlg.ShowDialog()==DialogResult.OK) {
-				textPathToExe.Text=dlg.SelectedPath;
+			OpenFileDialog fdlg=new OpenFileDialog();
+			fdlg.Title="Select a Service";
+			fdlg.InitialDirectory=Directory.GetCurrentDirectory();
+			fdlg.Filter="Executable files(*.exe)|*.exe";
+			fdlg.RestoreDirectory=true;
+			if(fdlg.ShowDialog()!=DialogResult.OK) {
+				return;				
 			}
+			textPathToExe.Text=fdlg.FileName;
 		}
 	}
 }
